requirements.txt
discord.py
python-dotenv
imaplib2
